requirements.txt
# Core AI & Agents
crewai>=0.28.0
crewai[tools]>=0.28.0

# Environment & Config
python-dotenv>=1.0.1

# Web & Data
beautifulsoup4>=4.12.3
requests>=2.31.0
usaddress>=0.5.10    # For high-accuracy US/CA address parsing

# Image Processing (ensure BICUBIC constant and modern Resampling)
pillow>=9.1.0,<11.0.0

# API Framework
flask>=3.0.3
flask-cors>=4.0.1

# Data Validation & OpenAI Client
pydantic>=2.7,<3.0
openai>=1.0.0,<2.0.0
